+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 - Handle brace blocks on commands properly.
 - Break parent and return `do` blocks when called from a `command` node.
+- Handle edge cases with `if` statements where there is no body of the if (so it can't be converted to a ternary).
 
 ## [0.2.0] - 2019-02-06
 ### Added